duro-telegram-bridge/requirements.txt
python-telegram-bot>=21.0
anthropic>=0.40.0
mcp>=1.0.0
